backend/requirements.txt
asgiref==3.7.2
certifi==2026.1.4
charset-normalizer==3.4.4
Django==4.2.7
django-cors-headers==4.3.1
django-jazzmin==3.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.11
pillow==12.1.0
PyJWT==2.10.1
python-decouple==3.8
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.6.3
